+++ Serendipity/src/main/java/org/shawn/games/Serendipity/AlphaBeta.java
@@ -2,8 +2,7 @@ package org.shawn.games.Serendipity;
 
 import java.util.*;
 
-import org.shawn.games.Serendipity.NNUE.NNUE;
-import org.shawn.games.Serendipity.NNUE.NNUE.NNUEAccumulator;
+import org.shawn.games.Serendipity.NNUE.*;
 
 import com.github.bhlangonijr.chesslib.*;
 import com.github.bhlangonijr.chesslib.move.*;
@@ -40,8 +39,7 @@ public class AlphaBeta
 	private int selDepth;
 
 	private NNUE network;
-	private NNUEAccumulator blackAccumulator;
-	private NNUEAccumulator whiteAccumulator;
+	private AccumulatorManager accumulators;
 
 	private class SearchState
 	{
@@ -134,8 +132,11 @@ public class AlphaBeta
 	public int evaluate(Board board)
 	{
 		int v = (Side.WHITE.equals(board.getSideToMove())
-				? NNUE.evaluate(network, whiteAccumulator, blackAccumulator, NNUE.chooseOutputBucket(board))
-				: NNUE.evaluate(network, blackAccumulator, whiteAccumulator, NNUE.chooseOutputBucket(board))) * 24;
+				? NNUE.evaluate(network, accumulators.getWhiteAccumulator(), accumulators.getBlackAccumulator(),
+						NNUE.chooseOutputBucket(board))
+				: NNUE.evaluate(network, accumulators.getBlackAccumulator(), accumulators.getWhiteAccumulator(),
+						NNUE.chooseOutputBucket(board)))
+				* 24;
 		return v;
 	}
 
@@ -161,380 +162,6 @@ public class AlphaBeta
 		});
 
 		return moves;
-	}
-
-	private void fullAccumulatorUpdate(Board board, Side side)
-	{
-		if (side.equals(Side.WHITE))
-		{
-			whiteAccumulator.reset();
-			whiteAccumulator.setBucketIndex(NNUE.chooseInputBucket(board.getKingSquare(Side.WHITE).ordinal()));
-
-			for (Square sq : Square.values())
-			{
-				if (!board.getPiece(sq).equals(Piece.NONE))
-				{
-					whiteAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.WHITE));
-				}
-			}
-		}
-		else
-		{
-			blackAccumulator.reset();
-			blackAccumulator
-					.setBucketIndex(NNUE.chooseInputBucket(board.getKingSquare(Side.BLACK).ordinal() ^ 0b111000));
-
-			for (Square sq : Square.values())
-			{
-				if (!board.getPiece(sq).equals(Piece.NONE))
-				{
-					blackAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.BLACK));
-				}
-			}
-		}
-	}
-
-	private void fullAccumulatorUpdate(Board board)
-	{
-		whiteAccumulator.reset();
-		blackAccumulator.reset();
-		whiteAccumulator.setBucketIndex(NNUE.chooseInputBucket(board.getKingSquare(Side.WHITE).ordinal()));
-		blackAccumulator.setBucketIndex(NNUE.chooseInputBucket(board.getKingSquare(Side.BLACK).ordinal() ^ 0b111000));
-
-		for (Square sq : Square.values())
-		{
-			if (!board.getPiece(sq).equals(Piece.NONE))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.WHITE));
-				blackAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.BLACK));
-			}
-		}
-	}
-
-	private void updateAccumulators(Board board, Move move, boolean undo)
-	{
-		if (board.getSideToMove().equals(Side.WHITE) && board.getPiece(move.getFrom()).equals(Piece.WHITE_KING)
-				&& NNUE.chooseInputBucket(move.getFrom().ordinal()) != NNUE.chooseInputBucket(move.getTo().ordinal()))
-		{
-			if (undo)
-			{
-				fullAccumulatorUpdate(board, Side.WHITE);
-			}
-			else
-			{
-				board.doMove(move);
-				fullAccumulatorUpdate(board, Side.WHITE);
-				board.undoMove();
-			}
-		}
-		else
-		{
-			updateWhiteAccumulators(board, move, undo);
-		}
-
-		if (board.getSideToMove().equals(Side.BLACK) && board.getPiece(move.getFrom()).equals(Piece.BLACK_KING)
-				&& NNUE.chooseInputBucket(move.getFrom().ordinal() ^ 0b111000) != NNUE
-						.chooseInputBucket(move.getTo().ordinal() ^ 0b111000))
-		{
-			if (undo)
-			{
-				fullAccumulatorUpdate(board, Side.BLACK);
-			}
-			else
-			{
-				board.doMove(move);
-				fullAccumulatorUpdate(board, Side.BLACK);
-				board.undoMove();
-			}
-		}
-		else
-		{
-			updateBlackAccumulators(board, move, undo);
-		}
-	}
-
-	private void updateWhiteAccumulators(Board board, Move move, boolean undo)
-	{
-		if (undo)
-		{
-			if (board.getContext().isKingSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-
-				return;
-			}
-
-			if (board.getContext().isQueenSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.QUEEN_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-
-				return;
-			}
-
-			whiteAccumulator.addFeature(NNUE.getIndex(move.getFrom(), board.getPiece(move.getFrom()), Side.WHITE));
-
-			if (move.getPromotion().equals(Piece.NONE))
-			{
-				whiteAccumulator
-						.subtractFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getFrom()), Side.WHITE));
-			}
-
-			else
-			{
-				whiteAccumulator.subtractFeature(NNUE.getIndex(move.getTo(), move.getPromotion(), Side.WHITE));
-			}
-
-			if (!board.getPiece(move.getTo()).equals(Piece.NONE))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getTo()), Side.WHITE));
-			}
-
-			else if (move.getTo().equals(board.getEnPassant())
-					&& board.getPiece(move.getFrom()).getPieceType().equals(PieceType.PAWN))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(board.getEnPassantTarget(),
-						board.getPiece(board.getEnPassantTarget()), Side.WHITE));
-			}
-		}
-
-		else
-		{
-			if (board.getContext().isKingSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-
-				return;
-			}
-
-			if (board.getContext().isQueenSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.QUEEN_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.WHITE));
-				whiteAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.WHITE));
-
-				return;
-			}
-
-			whiteAccumulator.subtractFeature(NNUE.getIndex(move.getFrom(), board.getPiece(move.getFrom()), Side.WHITE));
-
-			if (move.getPromotion().equals(Piece.NONE))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getFrom()), Side.WHITE));
-			}
-
-			else
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(move.getTo(), move.getPromotion(), Side.WHITE));
-			}
-
-			if (!board.getPiece(move.getTo()).equals(Piece.NONE))
-			{
-				whiteAccumulator.subtractFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getTo()), Side.WHITE));
-			}
-
-			else if (move.getTo().equals(board.getEnPassant())
-					&& board.getPiece(move.getFrom()).getPieceType().equals(PieceType.PAWN))
-			{
-				whiteAccumulator.subtractFeature(NNUE.getIndex(board.getEnPassantTarget(),
-						board.getPiece(board.getEnPassantTarget()), Side.WHITE));
-			}
-		}
-	}
-
-	private void updateBlackAccumulators(Board board, Move move, boolean undo)
-	{
-		if (undo)
-		{
-			if (board.getContext().isKingSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-
-				return;
-			}
-
-			if (board.getContext().isQueenSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.QUEEN_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-
-				return;
-			}
-
-			blackAccumulator.addFeature(NNUE.getIndex(move.getFrom(), board.getPiece(move.getFrom()), Side.BLACK));
-
-			if (move.getPromotion().equals(Piece.NONE))
-			{
-				blackAccumulator
-						.subtractFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getFrom()), Side.BLACK));
-			}
-
-			else
-			{
-				blackAccumulator.subtractFeature(NNUE.getIndex(move.getTo(), move.getPromotion(), Side.BLACK));
-			}
-
-			if (!board.getPiece(move.getTo()).equals(Piece.NONE))
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getTo()), Side.BLACK));
-			}
-
-			else if (move.getTo().equals(board.getEnPassant())
-					&& board.getPiece(move.getFrom()).getPieceType().equals(PieceType.PAWN))
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(board.getEnPassantTarget(),
-						board.getPiece(board.getEnPassantTarget()), Side.BLACK));
-			}
-		}
-
-		else
-		{
-			if (board.getPiece(move.getFrom()).getPieceType().equals(PieceType.KING))
-			{
-				board.doMove(move);
-				fullAccumulatorUpdate(board);
-				board.undoMove();
-				return;
-			}
-			if (board.getContext().isKingSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.KING_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-
-				return;
-			}
-
-			if (board.getContext().isQueenSideCastle(move)
-					&& (board.getCastleRight(board.getSideToMove()).equals(CastleRight.QUEEN_SIDE)
-							|| board.getCastleRight(board.getSideToMove()).equals(CastleRight.KING_AND_QUEEN_SIDE)))
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.addFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getTo(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getRookCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.ROOK), Side.BLACK));
-				blackAccumulator.subtractFeature(NNUE.getIndex(
-						board.getContext().getKingCastleMove(board.getSideToMove(), CastleRight.QUEEN_SIDE).getFrom(),
-						Piece.make(board.getSideToMove(), PieceType.KING), Side.BLACK));
-
-				return;
-			}
-
-			blackAccumulator.subtractFeature(NNUE.getIndex(move.getFrom(), board.getPiece(move.getFrom()), Side.BLACK));
-
-			if (move.getPromotion().equals(Piece.NONE))
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getFrom()), Side.BLACK));
-			}
-
-			else
-			{
-				blackAccumulator.addFeature(NNUE.getIndex(move.getTo(), move.getPromotion(), Side.BLACK));
-			}
-
-			if (!board.getPiece(move.getTo()).equals(Piece.NONE))
-			{
-				blackAccumulator.subtractFeature(NNUE.getIndex(move.getTo(), board.getPiece(move.getTo()), Side.BLACK));
-			}
-
-			else if (move.getTo().equals(board.getEnPassant())
-					&& board.getPiece(move.getFrom()).getPieceType().equals(PieceType.PAWN))
-			{
-				blackAccumulator.subtractFeature(NNUE.getIndex(board.getEnPassantTarget(),
-						board.getPiece(board.getEnPassantTarget()), Side.BLACK));
-			}
-		}
 	}
 
 	private int quiesce(Board board, int alpha, int beta, int ply) throws TimeOutException
@@ -626,13 +253,13 @@ public class AlphaBeta
 				continue;
 			}
 
-			updateAccumulators(board, move, false);
+			accumulators.updateAccumulators(board, move, false);
 			board.doMove(move);
 
 			int score = -quiesce(board, -beta, -alpha, ply + 1);
 
 			board.undoMove();
-			updateAccumulators(board, move, true);
+			accumulators.updateAccumulators(board, move, true);
 
 			bestScore = Math.max(bestScore, score);
 			alpha = Math.max(alpha, bestScore);
@@ -788,7 +415,8 @@ public class AlphaBeta
 					&& !(PieceType.PAWN.equals(board.getPiece(move.getFrom()).getPieceType())
 							&& move.getTo() == board.getEnPassant());
 
-			if (isQuiet && !isPV && !inCheck && depth <= 6 && ss.moveCount > 3 + depth * depth && alpha > -MATE_EVAL + 1024)
+			if (isQuiet && !isPV && !inCheck && depth <= 6 && ss.moveCount > 3 + depth * depth
+					&& alpha > -MATE_EVAL + 1024)
 			{
 				continue;
 			}
@@ -799,7 +427,7 @@ public class AlphaBeta
 				continue;
 			}
 
-			updateAccumulators(board, move, false);
+			accumulators.updateAccumulators(board, move, false);
 			board.doMove(move);
 
 			inCheck = board.isKingAttacked();
@@ -839,7 +467,7 @@ public class AlphaBeta
 			}
 
 			board.undoMove();
-			updateAccumulators(board, move, true);
+			accumulators.updateAccumulators(board, move, true);
 
 			if (thisMoveEval > bestValue)
 			{
@@ -919,24 +547,14 @@ public class AlphaBeta
 		this.nodesLimit = limits.getNodes();
 		this.timeManager = new TimeManager(limits.getTime(), limits.getIncrement(), limits.getMovesToGo(), 100,
 				board.getMoveCounter());
+		this.searchStack = newSearchStack();
+		this.accumulators = new AccumulatorManager(network, board);
+
 		for (int i = 0; i < 13; i++)
 		{
 			for (int j = 0; j < 65; j++)
 			{
 				history[i][j] /= 5;
-			}
-		}
-		this.whiteAccumulator = new NNUEAccumulator(network, NNUE.chooseInputBucket(board, Side.WHITE));
-		this.blackAccumulator = new NNUEAccumulator(network, NNUE.chooseInputBucket(board, Side.BLACK));
-		this.searchStack = newSearchStack();
-
-		// Initialize Accumulators
-		for (Square sq : Square.values())
-		{
-			if (!board.getPiece(sq).equals(Piece.NONE))
-			{
-				whiteAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.WHITE));
-				blackAccumulator.addFeature(NNUE.getIndex(sq, board.getPiece(sq), Side.BLACK));
 			}
 		}
